Move loop inside a function to reduce the number of calls
In a profile of scour on a given input file,
removeDefaultAttributeValue is one of the top 3 most frequent function
being called with 332288 calls (with minidom.getAttribute and
minidom.hasAttribute being the only ones with similar number of
calls).

The high number of calls occurs because removeDefaultAttributeValue
only removes a single attribute at a time.  Given it is always called
in a trivial for loop, we can just move the loop inside
removeDefaultAttributeValue to reduce the number of calls.  At the
same time, rename the function to better reflect that it works on
multiple attributes.

Signed-off-by: Niels Thykier <niels@thykier.net>