Rewrite findReferencedElements without recursion
A classic "Depth-First-Search via call stack recusion to
Depth-First-Search with stack" rewrite.  The only thing worth noting
is that the nodes with same parent are visited in opposite order than
originally.  While trivially fixable, none of the callers appear to rely
on the order for anything at all, so I have left that out.

Signed-off-by: Niels Thykier <niels@thykier.net>