Support multiple input files with -o being a directory
This change makes it possible for scour to consume multiple input
files in one command invocation.  E.g.

     $ scour file1.svg file2.svgz ... output-directory
     # xargs friendly variant
     $ scour -o output-directory file1.svg file2.svgz ...

This avoids most of the "startup" overhead in python and scour when
many files are being processed.  On about a 100 of (already scour'ed)
gnuplot svg graphs, this change provides an almost 40% speed up compared
to a shell alternative:

     # Original shell pipeline (~29s)
     # Note; for bash, rewriting this without the "basename"-call does
     # not seem to improve performance considerably.
     $ for FILE in input/[01]* ; do \
         python3 -m scour.scour "$FILE" output/"$(basename "$FILE")" > /dev/null ; \
       done
    # With this patch (~16s)
    $ python3 -m scour.scour input/[01]* output > /dev/null

Signed-off-by: Niels Thykier <niels@thykier.net>