Optimize removeDefaultAttributeValues
Avoid looping over DefaultAttribute(s) that are not relevant for a
given node.  This skips a lot of calls to removeDefaultAttributeValue
but more importantly, it avoids "node.nodeName not in attribute.elements"
line in removeDefaultAttributeValue.  As attribute.elements is a list, this
becomes expensive for "larger lists" (or in this case when there are a lot
of attributes).

This seems to remove about 1½-2 minutes of runtime (out of ~8) on the
1_42_polytope_7-cube.svg test case provided in #184.

Signed-off-by: Niels Thykier <niels@thykier.net>